requirements.txt
amqp==2.6.1
api==0.0.7
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
CacheControl==0.12.6
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.22.0
colorama==0.4.3
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
distlib==0.3.1
distro==1.5.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.1
django-cors-middleware==1.5.0
django-heroku==0.3.1
django-nose==1.4.7
django-rest-registration==0.5.6
django-richtextfield==1.6
django-sendgrid-webhook==0.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
flake8==3.8.3
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
importlib-metadata==1.7.0
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
nose==1.3.7
numpy==1.19.2
packaging==20.3
pandas==1.1.2
pep517==0.8.2
pipenv==2020.8.13
progress==1.5
psycopg2==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.3.1
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
retrying==1.3.3
selenium==3.141.0
sendgrid==6.4.6
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
toml==0.10.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
zipp==3.1.0
